Migrate ProofHelper to ProofsHelper.
